--- a/ui/raidboss/popup-text.ts
+++ b/ui/raidboss/popup-text.ts
@@ -718,7 +718,7 @@ export class PopupText {
           // time later.  This will clobber each time we
           // load this, but that's ok.
           trigger.filename = setFilename;
-          const id = trigger.id ?? trigger.toString();
+          const id = trigger.id;
 
           if (!isRegexTrigger(trigger) && !isNetRegexTrigger(trigger)) {
             console.error(`Trigger ${id}: has no regex property specified`);
@@ -762,7 +762,7 @@ export class PopupText {
               trigger.localNetRegex = Regexes.parse(localeNetRegex);
               orderedTriggers.push(trigger);
               found = true;
-            } else if (defaultNetRegex) {
+            } else if (defaultNetRegex !== undefined) {
               // simple netRegex trigger, need to build netRegex and translate
               if (defaultNetRegex instanceof RegExp) {
                 const trans = translateRegex(defaultNetRegex, this.parserLang, set.timelineReplace);
@@ -787,7 +787,7 @@ export class PopupText {
           }
 
           if (!found) {
-            console.error(`Trigger ${id}: missing regex and netRegex`);
+            console.error('Trigger ' + trigger.id + ': missing regex and netRegex');
             continue;
           }
         }