This is a (almost) type only change, it change the type of `TriggerSet` only. Our core file are using `LooseTriggerSet` and `LooseTrigger` so they are not affected.

This patch make `trigger.type` required for all cactbot built-in net triggers and merge `BaseTrigger` and `PartialNetRegexTrigger` to help tsc narrow trigger type.

about why merging types is explained here: https://github.com/quisquous/cactbot/issues/4508#issuecomment-1298080356

## result

`npx tsc` 35s -> 20s
`npx eslint .` 1m6.862s -> 40.713s
`npm test` 38s -> 20s
`npm run start` no improvement.

---

I know the team of swc is trying to port tsc to rust, wish them good luck…