@trim21 2022-11-01
trace after this patch:

```text
~/proj/cactbot (speedup-tsc-a-little) # time npm run tsc-no-emit -- --generateTrace ../cactbot-tracing && npx analyze-trace ../cactbot-tracing --force-millis 200

> cactbot@0.27.21 tsc-no-emit
> tsc --noEmit --generateTrace ../cactbot-tracing


real    0m20.240s
user    0m0.046s
sys     0m0.166s
Hot Spots
├─ Check file c:\users\trim21\proj\cactbot\node_modules\typescript\lib\lib.dom.d.ts (366ms)
└─ Check file c:\users\trim21\proj\cactbot\ui\raidboss\data\05-shb\eureka\delubrum_reginae_savage.ts (208ms)
   └─ Check variable declaration from (line 94, char 7) to (line 3712, char 2) (185ms)
      └─ Check expression from (line 94, char 38) to (line 3712, char 2) (169ms)
         └─ Check expression from (line 2601, char 20) to (line 3711, char 4) (126ms)

No duplicate packages found
```

` Check expression from (line 2601, char 20) to (line 3711, char 4) (126ms)` is a huge `timelineReplace`