@quisquous 2022-11-02
Thanks for the investigation and the fix.  That's really smart to combine the types like that to help out tsc.